+++ lib/gpui/event.ex
@@ -6,7 +6,15 @@ defmodule GPUI.Event do
   defstruct [:type, :window_id, :event, :value, attrs: %{}]
 
   @type type ::
-          :click | :change | :release | :search | :keydown | :keyup | :window_closed | atom()
+          :click
+          | :change
+          | :select
+          | :release
+          | :search
+          | :keydown
+          | :keyup
+          | :window_closed
+          | atom()
   @type t :: %__MODULE__{
           type: type(),
           window_id: pos_integer() | nil,